docs: More all demos to separate (unpublished) package
