+++ packages/demos/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "@looker/components-demos",
+  "license": "MIT",
+  "version": "0.0.1",
+  "main": "src/index.ts",
+  "sideEffects": false,
+  "private": true,
+  "devDependencies": {
+    "@looker/components": "*",
+    "@looker/components-date": "*",
+    "@looker/design-tokens": "*",
+    "@looker/icons": "*",
+    "lodash": "*",
+    "react": "^16.14.0",
+    "react-copy-to-clipboard": "^5.0.3",
+    "react-is": "^16.13.1",
+    "styled-components": "^5.3.0"
+  }
+}
