+++ packages/demos/src/components/LocaleTable.tsx
