+++ www/src/MDX/shortcodes.ts
@@ -24,9 +24,4 @@
 
  */
 
-export * from '../documentation/foundations/demos'
-export * from '../documentation/components/content/demos'
-export * from '../documentation/components/date/demos'
-export * from '../documentation/components/ImportMessage'
-export * from '../documentation/components/layout/demos'
-export * from '../components/'
+export * from '@looker/components-demos'
